auto uses correct first name letter profile picture and in-sync at everyplace that uses the profile pic.
Note: not sure how disambigution works, but when I upload the same csv to two accounts, where I only change the profile pic of one person in one of the account, it also changes the profile for the other account. not sure if this is  disambigution from recognizing it's the same linkedin link or same full name from our connection collection.